pyproject.toml
[project]
name = "full-stack-gen-ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]
